Become a hunter on the landscape

This function allows the user of the GMSE software to insert themselves as a hunter in the simulation, allowing them to cull some number of resources in a time step as observed by the agent whose ID is 2.

be_hunter(OBSERVATION, AGENT, RESOURCES, LAND, PARAS, view, times)

Arguments

OBSERVATION
The observation array produced by the observation function within GMSE

AGENT
The array of agents produced in the main gmse() function

RESOURCES
The resources array produced by the resource function within GMSE

LAND
The landscape array on which interactions between resources and agents occur

PARAS
The vector of parameters that hold global and dynamic parameter values used by GMSE

view
The distance within which agents are able to observe resources on the landscape

times
The number of times that resources are observed in the observation model of GMSE

Value

the_land A cols by rows landscape with randomly distributed cell types

Examples

# NOT RUN {
HUNT_OUTCOME <- be_hunter(OBSERVATION_r, AGENTS, RESOURCES, LANDSCAPE_r,
paras, agent_view, times_observe);
# }