Density estimator of resource abundance

Estimates population size using simulated data produced by the observation model of GMSE -- it assumes that the density of resources observed on the subset of the landscape sampled equals the density on the whole landscape.

dens_est(observation, paras, view = view, land = land)

Arguments

observation
The array of resource observations from the observation model, used to estimate abundance of resources

paras
The vector of parameters that hold global and dynamic parameter values used by GMSE

view
This parameter determines the distance around an agent's location within which it can observe resources.

land
The landscape array on which interactions between resources and agents occur

Value

The density estimator (which is also performed GMSE in the manager function) returns a list that includes resource population size estimates along with 95

Examples

# NOT RUN {
analysis <- dens_est(observation = obs_t, paras = paras, view = view,
land = land1);
# }