Landscape initialisation

Initialise the landscape of the G-MSE model.

make_landscape(
  model,
  rows,
  cols,
  cell_types = 1,
  cell_val_mn = 1,
  cell_val_sd = 0,
  cell_val_max = 1,
  cell_val_min = 0,
  layers = 3,
  ownership = FALSE,
  owners = 4,
  public_land = 0,
  ownership_var = 0
)

Arguments

model

The type of model being applied (Currently only individual-based -- i.e., 'agent-based' -- models are allowed)

rows

The dimension of the other side of the landscape (e.g., Longitude)

cols

The dimension of one side of the landscape (e.g., Latitude)

cell_types

Scalar or vector of all possible types of landscape cells

cell_val_mn

Mean cell value (e.g., defining crop output on a cell)

cell_val_sd

Standard devation of cell values on a landscape

cell_val_max

The maximum value of a cell

cell_val_min

The minimum value of a cell

layers

The number of layers in the 3D landscape (should usually be 3)

ownership

A TRUE or FALSE whether land should be owned by stakeholders

owners

The number of stakeholders in the model that own land

public_land

The proportion of landscape cells that are not owned

ownership_var

Does distribution of land vary among users? >=0, <1

Value

the_land A cols by rows landscape with randomly distributed cell types

Examples

land <- make_landscape(model = "IBM", rows = 10, cols = 10, cell_types = 1, 
cell_val_mn = 1, cell_val_sd =  0)